requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
lightgbm
xgboost
joblib
kagglehub
python-multipart
google-generativeai
python-dotenv
pillow
requests
streamlit
googletrans==4.0.0rc1
